ai-model/requirements.txt
tensorflow>=2.10.0
xgboost>=1.7.3
scikit-learn>=1.2.2
pandas>=1.5.3
numpy>=1.23.5
matplotlib>=3.7.1
seaborn>=0.12.2
joblib>=1.2.0
shap>=0.41.0
iota-sdk>=1.0.0
aiohttp>=3.8.4
pydantic>=1.10.7
fastapi>=0.95.1
uvicorn>=0.22.0
python-dotenv>=1.0.0
requests>=2.28.2
pyarrow>=11.0.0
